[no ci] started refactoring external resolver errors
