+++ external-crates/move/crates/move-package-alt/src/dependency/external.rs
@@ -7,36 +7,28 @@
 use std::{
     collections::BTreeMap,
     fmt::Debug,
-    iter::once,
     ops::Range,
-    path::{Path, PathBuf},
-    process::Stdio,
+    process::{ExitStatus, Stdio},
 };
 
-use anyhow::bail;
-use futures::future::{join_all, try_join_all};
+use futures::future::try_join_all;
 use itertools::{Itertools, izip};
-use serde::{
-    Deserialize, Serialize,
-    de::{MapAccess, Visitor},
-};
-use serde_spanned::Spanned;
-use tokio::{io::AsyncReadExt, process::Command};
-use tracing::{debug, info, warn};
+use serde::{Deserialize, Serialize};
+use thiserror::Error;
+use tokio::process::Command;
+use tracing::{debug, info};
 
 use crate::{
-    errors::{
-        self, FileHandle, Located, ManifestError, ManifestErrorKind, PackageError, PackageResult,
-        ResolverError,
-    },
+    errors::{FileHandle, Located, ManifestError, ManifestErrorKind},
     flavor::MoveFlavor,
     jsonrpc::Endpoint,
     package::{EnvironmentName, PackageName},
 };
 
-use super::{DependencySet, PinnedDependencyInfo, UnpinnedDependencyInfo, pin};
+use super::{DependencySet, UnpinnedDependencyInfo};
 
 pub type ResolverName = String;
+pub type ResolverResult<T> = Result<T, ResolverError>;
 
 pub const RESOLVE_ARG: &str = "--resolve-deps";
 pub const RESOLVE_METHOD: &str = "resolve";
@@ -51,16 +43,55 @@ pub struct ExternalDependency {
 
     /// the `<data>` in `{ r.<res> = <data> }`
     data: toml::Value,
+
+    /// The source file containing this dependency
+    file: FileHandle,
+
+    /// The range within the source file
+    span: Range<usize>,
+}
+
+#[derive(Error, Debug)]
+pub enum ResolverError {
+    #[error("I/O Error when running external resolver {resolver}")]
+    IoError {
+        resolver: ResolverName,
+
+        #[source]
+        source: std::io::Error,
+    },
+
+    /// This indicates that the resolver was faulty
+    #[error("{resolver} did not follow the external resolver protocol: {message}")]
+    BadResolver {
+        resolver: ResolverName,
+        message: String,
+    },
+
+    /// This indicates that the resolver returned a non-successful exit code
+    #[error("{resolver} returned error code: {code}")]
+    ResolverUnsuccessful {
+        resolver: ResolverName,
+        code: ExitStatus,
+    },
+
+    /// This indicates that the resolver executed successfully but returned an error
+    #[error("{resolver} couldn't resolve {dep} in {env_str}: {message}")]
+    ResolverFailed {
+        resolver: ResolverName,
+        dep: PackageName,
+        env_str: String,
+        message: String,
+    },
 }
 
 /// Convenience type for serializing/deserializing external deps
 #[derive(Serialize, Deserialize)]
 struct RField {
-    r: BTreeMap<String, toml::Value>,
+    r: Located<BTreeMap<String, toml::Value>>,
 }
 
 /// Requests from the package mananger to the external resolver
-// TODO: fix typo above
 #[derive(Serialize, Debug)]
 struct ResolveRequest<F: MoveFlavor> {
     #[serde(default)]
@@ -89,7 +120,7 @@ impl ExternalDependency {
     pub async fn resolve<F: MoveFlavor>(
         deps: &mut DependencySet<UnpinnedDependencyInfo<F>>,
         envs: &BTreeMap<EnvironmentName, F::EnvironmentID>,
-    ) -> PackageResult<()> {
+    ) -> ResolverResult<()> {
         // we explode [deps] first so that we know exactly which deps are needed for each env.
         deps.explode(envs.keys().cloned());
 
@@ -136,38 +167,91 @@ impl ExternalDependency {
     }
 }
 
-// TODO: CLEANUP
-// Explain this conversion
-impl TryFrom<RField> for ExternalDependency {
-    type Error = PackageError;
+impl ResolverError {
+    pub fn io_error(resolver: &ResolverName, source: std::io::Error) -> Self {
+        Self::IoError {
+            resolver: resolver.clone(),
+            source,
+        }
+    }
 
+    pub fn bad_resolver(resolver: &ResolverName, message: impl AsRef<str>) -> Self {
+        Self::BadResolver {
+            resolver: resolver.clone(),
+            message: message.as_ref().to_string(),
+        }
+    }
+
+    pub fn nonzero_exit(resolver: &ResolverName, code: ExitStatus) -> Self {
+        Self::ResolverUnsuccessful {
+            resolver: resolver.clone(),
+            code,
+        }
+    }
+
+    pub fn resolver_failed(
+        resolver: ResolverName,
+        dep: PackageName,
+        env: Option<EnvironmentName>,
+        message: String,
+    ) -> Self {
+        Self::ResolverFailed {
+            resolver,
+            dep,
+            message,
+            env_str: match env {
+                Some(env_name) => format!("environment {env_name}"),
+                None => "default environment".to_string(),
+            },
+        }
+    }
+}
+
+impl TryFrom<RField> for ExternalDependency {
+    type Error = ManifestError;
+
+    /// Convert from [RField] (`{r.<res> = <data>}`) to [ExternalDependency] (`{ res, data }`)
     fn try_from(value: RField) -> Result<Self, Self::Error> {
         debug!("try_from: {:?}", value.r);
-        if value.r.len() != 1 {
-            return Err(PackageError::Generic("TODO".to_string()));
-            //            return Err(PackageError::Manifest(ManifestError {
-            //                kind: ManifestErrorKind::BadExternalDependency,
-            //                span: Some(value.r.span()),
-            //                handle: value.r.file(),
-            //            }));
+        if value.r.as_ref().len() != 1 {
+            return Err(ManifestError {
+                kind: ManifestErrorKind::BadExternalDependency,
+                span: Some(value.r.span()),
+                handle: value.r.file(),
+            });
         }
+
+        let file = value.r.file();
+        let span = value.r.span();
 
         let (resolver, data) = value
             .r
+            .into_inner()
             .into_iter()
             .next()
             .expect("iterator of length 1 structure is nonempty");
 
-        Ok(Self { resolver, data })
+        Ok(Self {
+            resolver,
+            data,
+            file,
+            span,
+        })
     }
 }
 
 impl From<ExternalDependency> for RField {
+    /// Translate from [ExternalDependency] `{ res, data }` to [RField] `{r.<res> = data}`
     fn from(value: ExternalDependency) -> Self {
-        let ExternalDependency { resolver, data } = value;
+        let ExternalDependency {
+            resolver,
+            data,
+            file,
+            span,
+        } = value;
 
         RField {
-            r: BTreeMap::from([(resolver, data)]),
+            r: Located::new(BTreeMap::from([(resolver, data)]), file, span),
         }
     }
 }
@@ -178,7 +262,7 @@ impl From<ExternalDependency> for RField {
 async fn resolve_single<F: MoveFlavor>(
     resolver: ResolverName,
     requests: DependencySet<ResolveRequest<F>>,
-) -> PackageResult<DependencySet<UnpinnedDependencyInfo<F>>> {
+) -> ResolverResult<DependencySet<UnpinnedDependencyInfo<F>>> {
     let mut child = Command::new(&resolver)
         .arg(RESOLVE_ARG)
         .stdin(Stdio::piped())
@@ -206,7 +290,10 @@ async fn resolve_single<F: MoveFlavor>(
         .map_err(|e| ResolverError::bad_resolver(&resolver, e.to_string()));
 
     // dump standard error
-    let output = child.wait_with_output().await?;
+    let output = child
+        .wait_with_output()
+        .await
+        .map_err(|e| ResolverError::io_error(&resolver, e))?;
 
     if !output.stderr.is_empty() {
         info!(
@@ -219,7 +306,7 @@ async fn resolve_single<F: MoveFlavor>(
     }
 
     if !output.status.success() {
-        return Err(ResolverError::nonzero_exit(&resolver, output.status).into());
+        return Err(ResolverError::nonzero_exit(&resolver, output.status));
     }
 
     let result: DependencySet<UnpinnedDependencyInfo<F>> = izip!(envs, pkgs, resps?).collect();
@@ -230,8 +317,7 @@ async fn resolve_single<F: MoveFlavor>(
             return Err(ResolverError::bad_resolver(
                 &resolver,
                 "resolvers must return resolved dependencies",
-            )
-            .into());
+            ));
         }
     }
