--- external-crates/move/crates/move-package-alt/src/errors/git_error.rs
@@ -1,1 +1,0 @@
-
